uhuh5

+++ internal/app/repository/product_repository.go
@@ -16,7 +16,7 @@ type ProductRepository interface {
 	GetProductsByCategoryID(id uint, language string) ([]entity.Product, error)
 	GetAllProducts(language string) ([]entity.Product, error)
 	GetProductsWithPagination(limit int,  offset int) ([]entity.Product, error)
-	GetProductsByFilter(discount []uint, minPrice uint , maxPrice uint , subcategoryID []uint) ([]entity.Product , error)
+	GetProductsByFilter(discount []int, minPrice int , maxPrice int , subcategoryID []int) ([]entity.Product , error)
 	GetProductByID(id uint, language string) (*entity.Product, error)
 	CreateProduct(product *entity.Product) error
 	UpdateProduct(product *entity.Product, id uint) error
@@ -269,7 +269,7 @@ func (pr *productRepository) GetProductsByCategoryID(id uint, language string) (
 }
 
 
-func (pr *productRepository) GetProductsByFilter(discount []uint , minPrice uint , maxPrice uint , subcategoryID []uint) ([]entity.Product , error) {
+func (pr *productRepository) GetProductsByFilter(discount []int , minPrice int , maxPrice int , subcategoryID []int) ([]entity.Product , error) {
 
 	var products []entity.Product
 
@@ -278,19 +278,19 @@ func (pr *productRepository) GetProductsByFilter(discount []uint , minPrice uint
 	// Проверка на наличие значений для фильтрации
 	if len(subcategoryID) != 0 {
     for _ , value := range subcategoryID {
-		  query = query.Where("sub_category_id = ?", value)
+		  query = query.Where("sub_category_id = ?", uint(value))
     }  
   }
 	if minPrice != 0 {
-		query = query.Where("price >= ?", minPrice)
+		query = query.Where("price >= ?", uint(minPrice))
 	}
 	if maxPrice != 0 {
-		query = query.Where("price <= ?", maxPrice)
+		query = query.Where("price <= ?", uint(maxPrice))
 	}
 
 	if len(discount) != 0 {
     for _ , value := range discount {
-		  query = query.Where("discount = ?" , value)
+		  query = query.Where("discount = ?" , uint(value))
 	  }
   }
